pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ask-cbioportal"
version = "0.1.0"
description = "Natural language interface for querying cBioPortal cancer genomics data"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Craig Myles" }
]
keywords = ["cbioportal", "cancer", "genomics", "cli", "ai", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.32.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[project.scripts]
ask-cbioportal = "ask_cbioportal.cli:main"

[project.urls]
Homepage = "https://github.com/craigmyles/ask-cbioportal"
Repository = "https://github.com/craigmyles/ask-cbioportal"

[tool.hatch.build.targets.wheel]
packages = ["src/ask_cbioportal"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
